Add version suffix v1.0.0 to all default ClusterClass names
According to ClusterClass best practice, when we make any changes on it
or its reference templates, we 'd better make a copy of all and rename
them. Given multiple versions of the default ClusterClasses will coexist
in the same management cluster, we add a version suffix in their names
to identify them and avoid package conflicts.

There is no time to separate ClusterClass packages from the TKG meta
package in Glasgow, we decided to hard code version v1.0.0 for now and
will start to package them individually from Halifax.